+++ src/app/admin/components/index.ts
@@ -1,0 +1,3 @@
+export * from './products/products.component';
+export * from './manage-product/manage-product.component';
+export * from './orders/orders.component';