+++ src/app/cart/services/cart.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ICartItem } from '../models/cart-item.interface';
 import { IProductModel } from '../../products/models/product-model.interface';
+import { LocalStorageService } from '../../core/services';
 
 @Injectable({
   providedIn: 'root'
@@ -10,36 +11,57 @@ export class CartService {
   public totalSum = 0;
 
   get isEmptyCart(): boolean {
-    return this.cartProducts.length === 0;
+    return this.cartProducts ? this.cartProducts.length === 0 : true;
   }
 
-  private cartProducts: ICartItem[] = [];
+  public get cartProducts(): ICartItem[] | null {
+    return this.localStorageService.getItem('cart-products');
+  }
 
-  constructor() { }
+  public set cartProducts(value: ICartItem[] | null) {
+    if (value === null) {
+      this.localStorageService.removeItem('cart-products');
+    }
+    else {
+      this.localStorageService.setItem('cart-products', value);
+    }
+  }
+
+  constructor(
+    private localStorageService: LocalStorageService
+  ) {
+  }
 
   getProducts(): ICartItem[] {
-    return this.cartProducts;
+    return this.cartProducts ? this.cartProducts : [];
   }
 
   addProduct(product: IProductModel, count = 1): void {
-    const correspondCartItem = this.cartProducts.find(x => x.item.id === product.id);
+    const products = this.getProducts();
+    const correspondCartItem = products.find(x => x.item.id === product.id);
     if (correspondCartItem) {
       correspondCartItem.count += count;
-    } else {
-      this.cartProducts.push({
+    }
+    else {
+      products.push({
         count,
         item: product,
         name: product.name,
         price: product.price
       });
     }
+
+    this.cartProducts = products;
   }
 
   removeProduct(id: number): void {
-    const cartItemIndex = this.cartProducts.findIndex(x => x.item.id === id);
-    if (cartItemIndex >= 0){
-      this.cartProducts.splice(cartItemIndex, 1);
+    const products = this.getProducts();
+    const cartItemIndex = products.findIndex(x => x.item.id === id);
+    if (cartItemIndex >= 0) {
+      products.splice(cartItemIndex, 1);
     }
+
+    this.cartProducts = products;
   }
 
   increaseQuantity(id: number, value = 1): void {
@@ -55,22 +77,26 @@ export class CartService {
   }
 
   updateCartData(): void {
-    this.totalQuantity = this.cartProducts.reduce((previousValue, currentValue) => {
+    const products = this.getProducts();
+    this.totalQuantity = products.reduce((previousValue, currentValue) => {
       return previousValue + currentValue.count;
     }, 0);
 
-    this.totalSum = this.cartProducts.reduce((previousValue, currentValue) => {
+    this.totalSum = products.reduce((previousValue, currentValue) => {
       return previousValue + currentValue.item.price;
     }, 0);
   }
 
-  private changeQuantity(id: number, value: number): void {
-    const correspondCartItem = this.cartProducts.find(x => x.item.id === id);
+  changeQuantity(id: number, value: number): void {
+    const products = this.getProducts();
+    const correspondCartItem = products.find(x => x.item.id === id);
     if (correspondCartItem) {
       correspondCartItem.count = correspondCartItem.count + value;
       if (correspondCartItem.count <= 0) {
         this.removeProduct(id);
       }
     }
+
+    this.cartProducts = products;
   }
 }