+++ src/app/core/core.module.ts
@@ -1,9 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { GeneratorFactory } from './services/generator.factory';
-import { GeneratorService } from './services/generator.service';
-import { CONSTANTS_SERVICE } from './services/constants.service';
-import { LocalStorageService } from './services/local-storage.service';
+import { CONSTANTS_SERVICE, GeneratorService, LocalStorageService } from './services';
 import { generatedString } from './services/generated-string.injection-token';
 
 
@@ -14,6 +12,7 @@ import { generatedString } from './services/generated-string.injection-token';
   ],
   providers: [
     GeneratorService,
+    LocalStorageService,
     {
       provide: generatedString,
       useFactory: GeneratorFactory(10),
@@ -22,11 +21,8 @@ import { generatedString } from './services/generated-string.injection-token';
     {
       provide: CONSTANTS_SERVICE,
       useValue: CONSTANTS_SERVICE
-    },
-    {
-      provide: LocalStorageService,
-      useValue: LocalStorageService
     }
   ]
 })
-export class CoreModule { }
+export class CoreModule {
+}