+++ src/app/core/guards/auth.guard.ts
@@ -1,0 +1,55 @@
+import { Injectable } from '@angular/core';
+import {
+  CanActivate,
+  ActivatedRouteSnapshot,
+  RouterStateSnapshot,
+  UrlTree,
+  Router,
+  CanActivateChild,
+  NavigationExtras, Route, UrlSegment, CanLoad
+} from '@angular/router';
+import { Observable } from 'rxjs';
+import { AuthService } from '../services/auth.service';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthGuard implements CanActivate, CanActivateChild, CanLoad {
+  constructor(
+    private authService: AuthService,
+    private router: Router
+  ) {
+  }
+
+  canActivate(
+    route: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    const {url} = state;
+    return this.checkLogin(url);
+  }
+
+  canActivateChild(next: ActivatedRouteSnapshot, state: RouterStateSnapshot):
+    Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    const {url} = state;
+    return this.checkLogin(url);
+  }
+
+  canLoad(route: Route, segments: UrlSegment[]): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    const url = `/${route.path}`;
+    return this.checkLogin(url) as boolean;
+  }
+
+  private checkLogin(url: string): boolean | UrlTree {
+    if (this.authService.isLoggedIn) {
+      return true;
+    }
+    this.authService.redirectUrl = url;
+    const sessionId = 123456789;
+    const navigationExtras: NavigationExtras = {
+      queryParams: {sessionId},
+      fragment: 'anchor'
+    };
+    this.router.navigate(['/login'], navigationExtras);
+    return false;
+  }
+}