+++ src/app/core/services/local-storage.service.ts
@@ -1,16 +1,23 @@
 import { Injectable } from '@angular/core';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class LocalStorageService {
 
   constructor() { }
 
-  getItem(key: string): string | null {
-    return localStorage.getItem(key);
+  getItem(key: string): any | null {
+    const item = localStorage.getItem(key);
+    if (item) {
+      return JSON.parse(item);
+    } else {
+      return item;
+    }
   }
 
-  setItem(key: string, value: string): void {
-    localStorage.setItem(key, value);
+  setItem(key: string, value: any): void {
+    localStorage.setItem(key, JSON.stringify(value));
   }
 
   removeItem(key: string): void {