+++ src/app/products/products.module.ts
@@ -2,18 +2,18 @@ import { NgModule } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { SharedModule } from '../shared/shared.module';
-import { ProductsComponent } from './products.component';
 import { ProductsRoutingModule } from './products-routing.module';
+import { CoreModule } from '../core/core.module';
 
 @NgModule({
   declarations: [
     ProductsRoutingModule.components,
-    ProductsComponent,
   ],
   imports: [
     ProductsRoutingModule,
     MatCardModule,
     MatButtonModule,
+    CoreModule,
     SharedModule
   ]
 })